rename layer controls
