+++ napari/_qt/layer_controls/qt_image_controls.py
