+++ napari/_qt/layer_controls/qt_labels_controls.py
