+++ napari/_qt/layer_controls/qt_layer_controls_base.py
