+++ napari/_qt/layer_controls/qt_layer_controls_container.py
