+++ napari/_qt/layer_controls/qt_points_controls.py
