+++ napari/_qt/layer_controls/qt_shapes_controls.py
