+++ napari/_qt/layer_controls/qt_surface_controls.py
