+++ napari/_qt/layer_controls/qt_vectors_controls.py
